% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMeanTest.R
\name{tMeanTest}
\alias{tMeanTest}
\title{t-test for means and standard errors, assuming unpooled variance.}
\usage{
tMeanTest(mean.x, se.x, mean.y, se.y, alternative = "two.sided")
}
\arguments{
\item{mean.x}{Mean of sample of random variable X.}

\item{se.x}{Standard error of sample of random variable X.}

\item{mean.y}{Mean of sample of random variable Y.}

\item{se.y}{Standard error of sample of random variable Y.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater", or "less".}
}
\value{
List of the 't' statistic and boolean 'isSignificant',
which is calculated at the 95\% confidence level,
i.e. within 1.96 standard deviations of the mean.
}
\description{
t-test for means and standard errors, assuming unpooled variance.
}
\details{
Assumptions: The two samples are drawn from normal distributions.
The two samples do not need to have the same variance.
}
\note{
Standard error = (standard deviation ^ 2) / N, i.e. the square of
the standard deviation divided by the size of the sample.
}
